requirements.txt
asgiref==3.2.7
astroid==2.2.5
colorama==0.4.1
dj-database-url==0.5.0
Django==2.1.7
django-crispy-forms==1.9.0
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysql-connector-python==8.0.17
mysqlclient==1.4.6
pylint==2.3.1
PyMySQL==0.9.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.1
typed-ast==1.4.0
whitenoise==5.0.1
wrapt==1.11.2
